feat: scanner status
